update ExitDispute test
