+++ test/Dispute/ExitDispute.test.ts
@@ -17,7 +17,6 @@ import * as ethers from 'ethers'
 import {
   Address,
   Bytes,
-  BigNumber,
 } from '@cryptoeconomicslab/primitives'
 import EthCoder from '@cryptoeconomicslab/eth-coder'
 import { setupContext } from '@cryptoeconomicslab/context'
@@ -44,6 +43,7 @@ describe('ExitDispute', () => {
   let commitment: ethers.Contract
   let commitmentVerifier: ethers.Contract
   let mockCompiledPredicate: ethers.Contract
+  let spentChallengeValidator: ethers.Contract
 
   before(async () => {
     utils = await deployContract(wallet, Utils, [])
@@ -69,6 +69,7 @@ describe('ExitDispute', () => {
     disputeManager = await deployContract(wallet, DisputeManager, [
       utils.address
     ])
+    //spentChallengeValidator = await deployContract(wallet, SpentChallengeValidator, [])
     exitDispute = await deployContract(wallet, ExitDispute, [
       disputeManager.address,
       commitmentVerifier.address,
@@ -179,23 +180,16 @@ describe('ExitDispute', () => {
         gasLimit: 800000
       })
       // prepare challenge
-      let challengeInputs
-      if (challengeType === 'EXIT_SPENT_CHALLENGE') {
-        challengeInputs = [
-          Bytes.fromString(challengeType)
-        ]
-      } else {
-        challengeInputs = [
-          Bytes.fromString(challengeType)
-        ]
-      }
+      const challengeInputs = [
+        Bytes.fromString(challengeType)
+      ]
       const challengeWitness = [
         encodeStructable(firstBlockInfo.inclusionProof)
       ]
       return [inputs, challengeInputs, challengeWitness]
     }
     describe('succeed to exit challenge', () => {
-      it.only('create a new exit challenge(spent)', async () => {
+      it('create a new exit challenge(spent)', async () => {
         const [inputs, challengeInputs, challengeWitness] = await init()
         await mockCompiledPredicate.setDicideReturn(true)
         const transaction = support.ownershipTransaction(
@@ -212,7 +206,7 @@ describe('ExitDispute', () => {
             challengeInputs,
             challengeWitness,
             {
-              gasLimit: 800000
+              gasLimit: 900000
             }
           )
         ).to.emit(exitDispute, 'ExitChallenged')
@@ -267,6 +261,72 @@ describe('ExitDispute', () => {
             [],
             {
               gasLimit: 800000
+            }
+          )
+        ).to.be.reverted
+      })
+      it('token must be same', async () => {
+        const [inputs, challengeInputs, challengeWitness] = await init()
+        await mockCompiledPredicate.setDicideReturn(true)
+        const transaction = support.ownershipTransaction(
+          Address.from(ALICE_ADDRESS),
+          1000000,
+          0,
+          5,
+          Address.from(mockCompiledPredicate.address)
+        )
+        challengeInputs.push(EthCoder.encode(toTransactionStruct(transaction)))
+        await expect(
+          exitDispute.challenge(
+            inputs,
+            challengeInputs,
+            challengeWitness,
+            {
+              gasLimit: 900000
+            }
+          )
+        ).to.be.reverted
+      })
+      it('range must contain subrange', async () => {
+        const [inputs, challengeInputs, challengeWitness] = await init()
+        await mockCompiledPredicate.setDicideReturn(true)
+        const transaction = support.ownershipTransaction(
+          Address.from(BOB_ADDRESS),
+          1000000,
+          1,
+          5,
+          Address.from(mockCompiledPredicate.address)
+        )
+        challengeInputs.push(EthCoder.encode(toTransactionStruct(transaction)))
+        await expect(
+          exitDispute.challenge(
+            inputs,
+            challengeInputs,
+            challengeWitness,
+            {
+              gasLimit: 900000
+            }
+          )
+        ).to.be.reverted
+      })
+      it('State object decided to false', async () => {
+        const [inputs, challengeInputs, challengeWitness] = await init()
+        await mockCompiledPredicate.setDicideReturn(false)
+        const transaction = support.ownershipTransaction(
+          Address.from(BOB_ADDRESS),
+          1000000,
+          0,
+          5,
+          Address.from(mockCompiledPredicate.address)
+        )
+        challengeInputs.push(EthCoder.encode(toTransactionStruct(transaction)))
+        await expect(
+          exitDispute.challenge(
+            inputs,
+            challengeInputs,
+            challengeWitness,
+            {
+              gasLimit: 900000
             }
           )
         ).to.be.reverted
